Refactor to CLI
